♻️ refactor(web): 消除 OpenAPI hack，完善动态路由体系
